requirements.txt
absl-py==0.7.0
appnope==0.1.0
astor==0.7.1
astroid==1.6.3
autopep8==1.3.3
bleach==1.5.0
boto==2.49.0
boto3==1.9.174
botocore==1.12.174
certifi==2018.4.16
chardet==3.0.4
Click==7.0
coverage==4.4.2
cycler==0.10.0
Cython==0.27.3
decorator==4.1.2
docutils==0.14
enum34==1.1.6
Flask==1.0.3
gast==0.2.2
gensim==3.7.3
grpcio==1.18.0
h5py==2.8.0
html5lib==0.9999999
idna==2.7
imbalanced-learn==0.4.3
imblearn==0.0
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.11.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.11
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lazy-object-proxy==1.3.1
lxml==4.2.3
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==2.1.1
mccabe==0.6.1
mecab-python3==0.996.2
mock==3.0.5
nose==1.3.7
numpy==1.13.3
opencv-python==4.1.0.25
pandas==0.21.1
parso==0.1.0
pep8==1.7.1
pexpect==4.3.0
pickleshare==0.7.4
Pillow==6.0.0
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.5.2
pycodestyle==2.3.1
pydot==1.4.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.4
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.13
requests==2.19.1
s3transfer==0.2.1
scikit-learn==0.20.2
scipy==1.0.0
selenium==3.13.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
smart-open==1.8.4
tensorboard==1.13.0
tensorboard==1.13.0
tensorflow==1.13.1
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
